refactor(clawdstrike-logos): split verifier analysis module and normalize test fmt
Extract the formula-classification and atom-analysis free-function cluster
from verifier/verifier_core.rs (653 -> 355 lines) into a new sibling module
verifier/analysis.rs (308 lines), keeping verifier_core.rs to the
PolicyVerifier struct, its Default impl, and its inherent impl. Pure
relocation: the moved pub(crate) helpers (collect_atoms, classify_formula,
expected_action_types_for_policy, and the z3-gated counterexample helpers)
keep their exact visibility and #[cfg(feature = "z3")] attributes, and are
re-exported via 'pub(crate) use analysis::*;' so siblings and tests resolve
them unchanged. The now-empty 'pub(crate) use verifier_core::*;' glob gains
#[allow(unused_imports)], mirroring the existing load_time pattern.

Normalize verifier/tests formatting: cargo fmt de-indents the vestigial
one-level indentation in tests/mod.rs, and the include!'d theme files
(tests/{core,inheritance,witness_synthesis,load_time,z3}.rs) are de-indented
by hand (rustfmt does not follow include!). Test logic and the YAML raw-string
fixtures are unchanged (git diff -w is empty). cargo fmt --all -- --check now
passes.
